src/requirements.txt
python-dotenv
aiohttp
teams-ai
langchain
openai
langchain-openai
requests
langchain-chroma
bs4
sqlite3